Unverified completion
The app is complete, according to this project's tutorial, but I am not able to verify whether it works or not; the tutorial includes some deprecated code which is probably causing issues resulting in errors since the first run in the middle of the tutorial. I will revisit this project as soon as I hear back from the author of "100 Days with Swift", Paul Hudson, whom I asked about a possible update of the appropriate sections within this part of the course. Or when I will have some spare time and be skilful enough to update the deprecated code myself to work properly.